chore: move all related function into the engine module
