+++ wren-ai-service/src/utils.py
@@ -1,14 +1,11 @@
 import functools
 import logging
 import os
-import re
 import time
 from datetime import datetime
 from pathlib import Path
-from typing import Dict, List, Optional, Tuple
+from typing import Tuple
 
-import requests
-import sqlglot
 from dotenv import load_dotenv
 
 from src.core.provider import DocumentStoreProvider, LLMProvider
@@ -52,21 +49,6 @@ def setup_custom_logger(name, level=logging.INFO):
     return logger
 
 
-def clean_generation_result(result: str) -> str:
-    def _normalize_whitespace(s: str) -> str:
-        return re.sub(r"\s+", " ", s).strip()
-
-    return (
-        _normalize_whitespace(result)
-        .replace("\\n", " ")
-        .replace("```sql", "")
-        .replace('"""', "")
-        .replace("'''", "")
-        .replace("```", "")
-        .replace(";", "")
-    )
-
-
 def load_env_vars() -> str:
     load_dotenv(override=True)
 
@@ -76,101 +58,6 @@ def load_env_vars() -> str:
         load_dotenv(".env.prod", override=True)
 
     return "dev" if is_dev_env else "prod"
-
-
-def remove_limit_statement(sql: str) -> str:
-    pattern = r"\s*LIMIT\s+\d+(\s*;?\s*--.*|\s*;?\s*)$"
-    modified_sql = re.sub(pattern, "", sql, flags=re.IGNORECASE)
-
-    return modified_sql
-
-
-def add_quotes(sql: str) -> str:
-    logger.debug(f"Original SQL: {sql}")
-
-    quoted_sql = sqlglot.transpile(sql, read="trino", identify=True)[0]
-
-    logger.debug(f"Quoted SQL: {quoted_sql}")
-
-    return quoted_sql
-
-
-def classify_invalid_generation_results(
-    api_endpoint: str,
-    generation_results: List[Dict[str, str]],
-) -> List[Optional[Dict[str, str]]]:
-    valid_generation_results = []
-    invalid_generation_results = []
-
-    for generation_result in generation_results:
-        quoted_sql = add_quotes(generation_result["sql"])
-
-        response = requests.get(
-            f"{api_endpoint}/v1/mdl/dry-run",
-            json={
-                "sql": remove_limit_statement(quoted_sql),
-                "limit": 1,
-            },
-        )
-        if response.status_code == 200:
-            valid_generation_results.append(
-                {
-                    "sql": quoted_sql,
-                    "summary": generation_result["summary"],
-                }
-            )
-        else:
-            invalid_generation_results.append(
-                {
-                    "sql": quoted_sql,
-                    "summary": generation_result["summary"],
-                    "error": response.json()["message"],
-                }
-            )
-
-    return valid_generation_results, invalid_generation_results
-
-
-def check_if_sql_executable(
-    api_endpoint: str,
-    sql: str,
-):
-    response = requests.get(
-        f"{api_endpoint}/v1/mdl/dry-run",
-        json={
-            "sql": remove_limit_statement(add_quotes(sql)),
-            "limit": 1,
-        },
-    )
-
-    if response.status_code != 200:
-        logger.debug(f"SQL is not executable: {response.json()}")
-
-    return True if response.status_code == 200 else False
-
-
-def remove_duplicates(dicts):
-    """
-    Removes duplicates from a list of dictionaries based on 'sql' and 'summary' fields.
-
-    Args:
-    dicts (list of dict): The list of dictionaries to be deduplicated.
-
-    Returns:
-    list of dict: A list of dictionaries after removing duplicates.
-    """
-    # Convert each dictionary to a tuple of (sql, summary) to make them hashable
-    seen = set()
-    unique_dicts = []
-    for d in dicts:
-        identifier = (
-            d["sql"],
-            d["summary"],
-        )  # This assumes 'sql' and 'summary' always exist
-        if identifier not in seen:
-            seen.add(identifier)
-            unique_dicts.append(d)
-    return unique_dicts
 
 
 def init_providers() -> Tuple[LLMProvider, DocumentStoreProvider]:
